feat: add end point controller and service of learn-service

+++ services/learn-service/src/main/java/com/phuc/learn_service/repository/ToeicPartRepository.java
@@ -3,5 +3,5 @@ package com.phuc.learn_service.repository;
 import com.phuc.learn_service.entity.ToeicPart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToeicPartRepository extends JpaRepository<ToeicPart, Long> {
+public interface ToeicPartRepository extends JpaRepository<ToeicPart, Integer> {
 }
